move latlon to xy poly function to coord systems
